dockerfiles: use docker cache for dependencies
With the layer change introduced the download of labgrid python dependencies is
now the major bottleneck which is executed for each image.

Enable caching for this as well.

Note that since pip install has no "install dependencies only" option
for pyproject.toml we use a script to generate a temporary
requirements.txt.

To give a rough benchmark figure the build time for incremental builds on a fairly weak 4 core
i3 @2.2GHz with mediocre network connection went from ~120s to ~16s (~5.5s
for each image)